修复cl warning

+++ include/render_graph/frontend/pass_node.hpp
@@ -48,7 +48,7 @@ public:
     void foreach_textures(eastl::function<void(TextureNode*, TextureEdge*)>);
     inline uint32_t textures_count() const
     {
-        return in_texture_edges.size() + out_texture_edges.size() + inout_texture_edges.size();
+        return (uint32_t)(in_texture_edges.size() + out_texture_edges.size() + inout_texture_edges.size());
     }
 
     gsl::span<BufferReadEdge*> buf_read_edges();
@@ -57,7 +57,7 @@ public:
     void foreach_buffers(eastl::function<void(BufferNode*, BufferEdge*)>);
     inline uint32_t buffers_count() const
     {
-        return in_buffer_edges.size() + out_buffer_edges.size() + ppl_buffer_edges.size();
+        return (uint32_t)(in_buffer_edges.size() + out_buffer_edges.size() + ppl_buffer_edges.size());
     }
 
     const EPassType pass_type = EPassType::None;
